// src/types/payroll/PayrollEmployeeResponse.ts
export interface PayrollEmployeeResponse {
  id: string | null;
  userId: string | null;
  userProfileId: string;
  personnelCode: string;
  personnelName: string;
  departmentName: string | null;
  basicSalary: number;
  targetSalary: number;
  targetThreshold: number;
  commission: number;
  parkingAllowance: number;
  fuelAllowance: number;
  phoneAllowance: number;
  seniorityAllowance: number;
  performanceAttitudeAllowance: number;
  punctualityDisciplineAllowance: number;
  outstandingAllowance: number;
  insuranceBalance: number;
  usaOfficeAllowance: number;
  hotBonus: number;
  monthlyBonus: number;
  managementAllowance: number;
  businessTripFee: number;
  mealAllowance: number;
  jobAllowance: number;
  clientEntertainment: number;
  socialInsuranceSalary: number;
  companySocialInsurance: number;
  companyHealthInsurance: number;
  companyUnemploymentInsurance: number;
  employeeSocialInsurance: number;
  employeeHealthInsurance: number;
  employeeUnemploymentInsurance: number;
  workdaySalary: number;
  taxableIncome: number;
  personalDeduction: number;
  dependentCount: number;
  dependentTaxDeductionAmount: number;
  dependentDeduction: number;
  assessableIncome: number;
  personalIncomeTax: number;
  netSalary: number;
  bankTransfer: number;
  cashPayment: number;
  salaryNote: string | null;
  standardWorkdays: number;
  actualWorkdays: number;
}